+++ components/Card/Card.styled.js
@@ -37,3 +37,17 @@ export const ContentStyle = styled.div`
   flex-direction: column;
   margin-left: 15rem;
 `;
+
+export const SearchContainer = styled.div`
+  display: flex;
+  justify-content: center;
+  margin-top: 2rem;
+  margin-bottom: 2rem;
+
+  input {
+    width: 33%;
+    height: 2rem;
+    border-radius: 20px;
+    text-align: center;
+  }
+`;
